Can now add the arrival time at the task

+++ utilities/include/mrta_utilities/mrta_solution.h
@@ -15,9 +15,9 @@ namespace MrtaSolution
     // "Model not feasible"
     std::string result_description;
 
-    float maximum_robot_schedule;
+    double maximum_robot_schedule;
 
-    float sum_of_all_robot_schedules;
+    double sum_of_all_robot_schedules;
 
     std::string solver_runtime;
   };
@@ -26,7 +26,7 @@ namespace MrtaSolution
     std::string robot_id;
 
     std::vector<std::string> task_attendance_sequence;
-    std::map<std::string, float> task_arrival_time_map;
+    std::map<std::string, double> task_arrival_time_map;
   };
 
   struct CompleteSolution
